ai_models/deep_learning_models.py
import tensorflow as tf
from tensorflow.keras.models import Sequential, Model
from tensorflow.keras.layers import Dense, LSTM, Conv2D, MaxPooling2D, Flatten, Dropout, Input
import numpy as np
import pandas as pd
from sklearn.preprocessing import StandardScaler
import joblib

class DeepLearningModels:
    def __init__(self):
        self.models = {}
        self.scalers = {}
    
    def create_personnel_behavior_model(self):
        """Neural network for complex personnel behavior pattern recognition"""
        model = Sequential([
            Dense(128, activation='relu', input_shape=(20,)),
            Dropout(0.3),
            Dense(64, activation='relu'),
            Dropout(0.2),
            Dense(32, activation='relu'),
            Dense(3, activation='softmax')  # High/Medium/Low risk
        ])
        model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])
        return model
    
    def create_mission_readiness_lstm(self):
        """LSTM for time-series mission readiness prediction"""
        model = Sequential([
            LSTM(50, return_sequences=True, input_shape=(30, 10)),
            Dropout(0.2),
            LSTM(50, return_sequences=False),
            Dropout(0.2),
            Dense(25),
            Dense(1, activation='sigmoid')
        ])
        model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['accuracy'])
        return model
    
    def create_facial_recognition_cnn(self):
        """CNN for facial recognition attendance system"""
        model = Sequential([
            Conv2D(32, (3, 3), activation='relu', input_shape=(128, 128, 3)),
            MaxPooling2D(2, 2),
            Conv2D(64, (3, 3), activation='relu'),
            MaxPooling2D(2, 2),
            Conv2D(128, (3, 3), activation='relu'),
            MaxPooling2D(2, 2),
            Flatten(),
            Dense(512, activation='relu'),
            Dropout(0.5),
            Dense(1000, activation='softmax')  # 1000 personnel IDs
        ])
        model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])
        return model
    
    def train_models(self):
        """Train all deep learning models with synthetic data"""
        # Generate synthetic training data
        X_behavior = np.random.rand(5000, 20)
        y_behavior = tf.keras.utils.to_categorical(np.random.randint(0, 3, 5000), 3)
        
        X_lstm = np.random.rand(3000, 30, 10)
        y_lstm = np.random.randint(0, 2, 3000)
        
        X_face = np.random.rand(10000, 128, 128, 3)
        y_face = tf.keras.utils.to_categorical(np.random.randint(0, 1000, 10000), 1000)
        
        # Train behavior model
        behavior_model = self.create_personnel_behavior_model()
        behavior_model.fit(X_behavior, y_behavior, epochs=10, batch_size=32, validation_split=0.2, verbose=0)
        self.models['behavior'] = behavior_model
        
        # Train LSTM model
        lstm_model = self.create_mission_readiness_lstm()
        lstm_model.fit(X_lstm, y_lstm, epochs=10, batch_size=32, validation_split=0.2, verbose=0)
        self.models['mission_lstm'] = lstm_model
        
        # Train CNN model (simplified for demo)
        cnn_model = self.create_facial_recognition_cnn()
        cnn_model.fit(X_face, y_face, epochs=5, batch_size=16, validation_split=0.2, verbose=0)
        self.models['facial_recognition'] = cnn_model
        
        print("✅ Deep learning models trained successfully")
        return self.models
    
    def predict_behavior_pattern(self, personnel_data):
        """Predict personnel behavior patterns"""
        if 'behavior' not in self.models:
            return {'risk_level': 'Medium', 'confidence': 0.75}
        
        # Simulate prediction
        prediction = np.random.choice(['Low', 'Medium', 'High'], p=[0.6, 0.3, 0.1])
        confidence = np.random.uniform(0.7, 0.95)
        
        return {
            'risk_level': prediction,
            'confidence': confidence,
            'factors': ['Performance trends', 'Training completion', 'Peer interactions']
        }
    
    def predict_mission_readiness_sequence(self, historical_data):
        """Predict mission readiness over time"""
        if 'mission_lstm' not in self.models:
            return {'readiness_trend': [0.85, 0.87, 0.89, 0.91, 0.88]}
        
        # Simulate LSTM prediction
        trend = [0.85 + 0.02 * i + np.random.normal(0, 0.02) for i in range(5)]
        return {
            'readiness_trend': trend,
            'predicted_peak': max(trend),
            'risk_periods': [i for i, val in enumerate(trend) if val < 0.8]
        }
    
    def recognize_face(self, image_data):
        """Facial recognition for attendance"""
        if 'facial_recognition' not in self.models:
            return {'personnel_id': 'IAF001234', 'confidence': 0.92}
        
        # Simulate face recognition
        personnel_id = f"IAF{np.random.randint(100000, 999999)}"
        confidence = np.random.uniform(0.85, 0.98)
        
        return {
            'personnel_id': personnel_id,
            'confidence': confidence,
            'timestamp': pd.Timestamp.now().isoformat(),
            'location': 'Main Gate'
        }

if __name__ == "__main__":
    dl_models = DeepLearningModels()
    trained_models = dl_models.train_models()
    
    # Save models
    for name, model in trained_models.items():
        model.save(f'ai_models/{name}_model.h5')
    
    print("🧠 Deep Learning Models Ready!")