requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
faker==20.1.0
python-dateutil==2.8.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Deep Learning
tensorflow==2.15.0
keras==2.15.0
torch==2.1.1
torchvision==0.16.1

# Computer Vision
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
Pillow==10.1.0

# Natural Language Processing
speechrecognition==3.10.0
pyttsx3==2.90
googletrans==4.0.0rc1
nltk==3.8.1
textblob==0.17.1

# Audio Processing
pyaudio==0.2.11
wave==0.0.2
librosa==0.10.1

# Database
sqlite3
firebase-admin==6.2.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0